+++ doc/CHANGELOG.md
@@ -1,3 +1,12 @@
+27/9/2015
+---------
+ - Eliminate librerie non usate
+ - Migliorato codice
+ - Migliorata documentazione
+ - Nuovi metodi
+ - Nuovo oggetto Profile
+ - Risolto bug con il metodo Group.members()
+
 24/9/2015
 ---------
  - Prima versione